re locate utils into utils/
