+++ src/utils/slugify.ts
@@ -1,28 +1,4 @@
-import { type ClassValue, clsx } from "clsx";
 import crypto from "crypto";
-import { twMerge } from "tailwind-merge";
-import { z } from "zod";
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
-}
-
-/**
- * Truncates the given text string to the specified limit, appending '...' if it exceeds the limit.
- */
-export function truncateText(str: string, limit = 100) {
-  // if we don't remove code blocks from markdown it will overflow
-  const withoutTripleBackticks = str.replace(/```[\s\S]*?```/g, "");
-
-  return withoutTripleBackticks.length > limit
-    ? withoutTripleBackticks.slice(0, limit) + "..."
-    : withoutTripleBackticks + "...";
-}
-
-/**
- * zod schema for validating urls
- */
-export const urlSchema = z.string().url();
 
 /**
  * Slugify
