+++ src/utils/zodSchemas.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+/**
+ * zod schema for validating urls
+ */
+export const urlSchema = z.string().url();
